[Modift] 구조 변경 - 폴더구조 생성 및 파일 분리
