+++ server-ssr.js
@@ -1,12 +1,12 @@
 import express from 'express';
-import { App } from './src-ssr/component.js';
-import { serverRender } from './src-ssr/main-ssr.js';
+import { App } from './src-ssr/App.js';
+import { serverRender } from './src-ssr/common/utils/serverRender.js';
 import { store } from './src-ssr/store.js';
 
 const app = express();
 
 app.use(express.json()); // Server, Client 상태 동기화를 위함
-app.use('/src-ssr', express.static('./src-ssr')); //static file import등 처리를 위함
+app.use('/public', express.static('./src-ssr/public')); //static file import등 처리를 위함
 
 app.put('/api/state', (req, res) => {
   store.hydration(req.body);
