+++ src-ssr/App.js
@@ -1,0 +1,17 @@
+import { NavigationBar } from './common/components/csr/NavigationBar.js';
+import { Router } from './router.js';
+
+export const App = ({ req }) => {
+  const path = req.path;
+
+  return /* html */ `
+    <header>
+      <a href="/">SSR Blog!!</a>
+      ${NavigationBar()}
+    </header>
+    ${Router(path)}
+    <footer>
+      Test SSR Blog with SSR Sample 
+    </footer>
+  `;
+};
