+++ src-ssr/common/components/Button.js
@@ -1,0 +1,5 @@
+export const Button = ({ text }) => {
+  return /* html */ `
+    <button>${text}</button>
+  `;
+};
