+++ src-ssr/common/components/csr/NavigationBar.js
@@ -1,0 +1,14 @@
+export const NavigationBar = () => {
+  return /* html */ `
+    <nav>
+      <ul>
+        <li>
+          <a href="/">Home</a>
+        </li>
+        <li>
+          <a href="/todo-list">Todo List</a>
+        </li>
+      </ul>
+    </nav>
+  `;
+};
