+++ src-ssr/common/utils/serverRender.js
@@ -5,7 +5,7 @@ export const serverRender = (RootComponent, serverState) => {
       <head>
         <meta charset="UTF-8">
         <title>Cloud Sanghun Server Side render Blog</title>
-        <link rel="preload stylesheet" href="../src-ssr/ssr.css" as="style"></link>
+        <link rel="preload stylesheet" href="../../public/ssr.css" as="style"></link>
         <script>
           window.state = ${JSON.stringify(serverState)}
         </script>
@@ -14,8 +14,7 @@ export const serverRender = (RootComponent, serverState) => {
         <div id="app">${RootComponent}</div>
 
         <!-- 브라우저에서의 CSR을 위한 Script 태그-->
-        <script src="../src-ssr/main-csr.js
-        " type="module"></script>
+        <script src="../../public/clientRender.js" type="module"></script>
       </body>
     </html>
   `;
