+++ src-ssr/pages/Home.js
@@ -1,0 +1,5 @@
+export const Home = () => {
+  return /* html */ `
+    <div>Home 입니다!</div>
+  `;
+};
